+++ terraform/slurm_cluster/examples/slurm_cluster/hybrid/full/variables.tf
@@ -217,12 +217,14 @@ Variables map to:
 EOD
   type = object({
     google_app_cred_path = string
+    slurm_control_host   = string
     slurm_bin_dir        = string
     slurm_log_dir        = string
     output_dir           = string
   })
   default = {
     google_app_cred_path = null
+    slurm_control_host   = null
     slurm_bin_dir        = "/usr/local/bin"
     slurm_log_dir        = "/var/log/slurm"
     output_dir           = "."