+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## \[5.0.3\]
 
+- Allow configuring controller hostname for hybrid deployments.
+
 ## \[5.0.2\]
 
 - Fix applying enable_bigquery_load to an existing cluster